backend/pyproject.toml
[project]
name = "backend"
version = "0.0.0"
description = "FastAPI backend for MAX Agentic Cookbook"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "openai>=2.6.0",
]

[project.scripts]
dev = "src.cli:dev"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = []
